+++ src/constants.ts
@@ -1,2 +1,3 @@
 export const VINMONOPOLET_SEARCH_URL =
-  "https://www.vinmonopolet.no/vmpws/v2/vmp/search/content";
+  "https://www.vinmonopolet.no/vmpws/v2/vmp/search";
+export const VINMONOPOLET_PRODUCT_URL = "https://www.vinmonopolet.no/p/";