+++ src/index.ts
@@ -3,7 +3,7 @@ export { default as Facet } from "./models/Facet";
 export { default as FacetValue } from "./models/FacetValue";
 export * as FacetCategory from "./models/FacetCategories";
 export { default as FoodPairing } from "./models/FoodPairing";
-export { default as Pagination } from "./models/Pagination";
+export { default as GetProductsPagination } from "./models/Pagination";
 export {
   default as BaseProduct,
   PopulatedProduct,
@@ -14,16 +14,12 @@ export { default as ProductStatus } from "./models/ProductStatus";
 export { default as RawMaterial } from "./models/RawMaterial";
 export { default as PopulatedStore, BaseStore } from "./models/Store";
 
-export {
-  default as getProducts,
-  getProductCount,
-} from "./retrievers/getProducts";
+export { default as getProducts } from "./retrievers/getProducts";
 export { default as getProduct } from "./retrievers/getProduct";
 export { default as getFacets } from "./retrievers/getFacets";
 export { default as getProductByBarcode } from "./retrievers/getProductByBarcode";
 export { default as getProductsByStore } from "./retrievers/getProductsByStore";
 export { default as getProductsByIds } from "./retrievers/getProductsById";
 export { default as getAllStores, searchStores } from "./retrievers/getStores";
-export { default as searchProducts } from "./retrievers/searchProducts";
 export { default as getStore } from "./retrievers/getStore";
 export { default as getProductReleases } from "./retrievers/getProductReleases";